requirements.txt
asn1crypto==0.22.0
bcrypt==3.1.3
cryptography>=2.3
django-split-settings==0.2.5
enum34==1.1.6
Fabric==1.13.2
idna==2.5
paramiko>=2.2.4
six==1.10.0